Code clean up, remove unused imports
